Storeをモジュールに切り分け
