+++ client/src/router/index.js
@@ -56,8 +56,8 @@ router.beforeEach(async (to, from, next) => {
     return next();
   }
 
-  await store.dispatch('checkAuthenticated');
-  const { isAuthenticated } = store.getters;
+  await store.dispatch('user/checkAuthenticated');
+  const isAuthenticated = store.getters['user/isAuthenticated'];
 
   // isPublic でない場合(=認証が必要な場合)、かつ、ログインしていない場合
   if (!to.meta.isPublic && !isAuthenticated) {
